refactor: 파일명 변경 및 리뷰 모아보는 모달 추가
[#412]
