+++ frontend/src/components/ui/StationDetailsWindow/StationDetailsView.tsx
@@ -15,7 +15,7 @@ import ChargerCard from '@ui/StationDetailsWindow/ChargerCard';
 import StationInformation from '@ui/StationDetailsWindow/StationInformation';
 import ChargerReportConfirmation from '@ui/StationDetailsWindow/reports/ChargerReportConfirmation';
 import StationReportPreConfirmation from '@ui/StationDetailsWindow/reports/StationReportPreConfirmation';
-import ReviewView from '@ui/StationDetailsWindow/reviews/ReviewView';
+import ReviewPreview from '@ui/StationDetailsWindow/reviews/ReviewPreview';
 
 import type { StationDetails } from '@type';
 
@@ -83,7 +83,7 @@ const StationDetailsView = ({ station }: StationDetailsViewProps) => {
       <Box my={3}>
         <CongestionStatistics />
       </Box>
-      <ReviewView stationId={stationId} />
+      <ReviewPreview stationId={stationId} />
     </Box>
   );
 };
