+++ frontend/src/components/ui/StationDetailsWindow/reviews/ReviewList.tsx
@@ -1,0 +1,5 @@
+const ReviewList = () => {
+  return <></>;
+};
+
+export default ReviewList;
